requirements.txt
numpy==1.22.4
pandas==1.5.3
seaborn==0.12.2
matplotlib==3.6.3
ydata-profiling==4.1.2
mrmr-selection==0.2.6
scikit-learn==1.2.2
xgboost==1.7.5
lightgbm==3.3.5
catboost==1.1.1
streamlit==1.21.0
pyarrow==9.0.0
pyyaml==6.0
pylint==2.17.2
pytest==7.2.2
-e .